refactor: move 'locking.now' to 'util.now'
Rework to avoid warning spew for 'datetime.datetime.utcnow'.

Suppress other warning spew (chameleon / AST, pkg_resources, etc.)
