+++ CHANGES.txt
@@ -1,3 +1,13 @@
+1.0 (unreleased)
+================
+
+- Move 'substanced.locking.now' to 'substanced.util.now', and rework to avoid
+  warning spew for 'datetime.datetime.utcnow'.
+ 
+- Suppress other warning spew from upstreams in newer Python versions
+  ('chameleon' / AST, 'pkg_resources', etc.)
+
+
 1.0b1 (2024-11-27)
 ==================
