+++ pytest.ini
@@ -3,5 +3,9 @@ addopts = -l --strict
 norecursedirs = lib include .tox .git
 python_files = test_*.py tests.py
 filterwarnings =
+     ignore::DeprecationWarning:chameleon
      ignore::DeprecationWarning:pkg_resources
+     ignore:.*pkg_resources:DeprecationWarning
+     ignore:.*Attribute . is deprecated:DeprecationWarning
+     ignore::DeprecationWarning:webob
      ignore::DeprecationWarning:zodburi
